+++ jason/src/transport/mod.rs
@@ -1,21 +1,21 @@
-pub mod protocol;
+mod pinger;
 mod websocket;
+
+pub mod protocol;
 
 use futures::sync::mpsc::UnboundedSender;
 use js_sys::Date;
-use wasm_bindgen::{prelude::*, JsCast};
-use web_sys::console;
 
-use std::{cell::RefCell, convert::TryFrom, rc::Rc, vec};
+use std::{cell::RefCell, rc::Rc, vec};
 
-use crate::utils::{window, IntervalHandle, WasmErr};
+use crate::{transport::protocol::{InMsg, OutMsg}, utils::WasmErr};
 
 use self::{
-    protocol::{Command, Event as MedeaEvent, Heartbeat},
-    websocket::{WebSocket},
+    pinger::Pinger,
+    protocol::{Command, Event as MedeaEvent},
+    websocket::WebSocket,
 };
-use crate::transport::protocol::InMsg;
-use web_sys::Event;
+
 
 // TODO:
 // 1. Reconnect.
@@ -58,7 +58,8 @@ impl Transport {
         let subs_rc = Rc::clone(&self.subs);
         socket_ref.on_message(move |msg: Result<InMsg, WasmErr>| {
             match msg {
-                Ok(InMsg::Pong(num)) => {
+                Ok(InMsg::Pong(_num)) => {
+                    // TODO: detect no pings
                     pinger_rc.set_pong_at(Date::now());
                 }
                 Ok(InMsg::Event(event)) => {
@@ -76,7 +77,7 @@ impl Transport {
                     err.log_err();
                 }
             }
-        });
+        })?;
 
         let socket_rc = Rc::clone(&socket);
         let pinger_rc: Rc<Pinger> = Rc::clone(&self.pinger);
@@ -84,12 +85,11 @@ impl Transport {
             socket_rc.borrow_mut().take();
             pinger_rc.stop();
 
-            //TODO: reconnect on disconnect
+            // TODO: reconnect on disconnect
             match msg {
-                CloseMsg::Normal(msg) => {},
-                CloseMsg::Disconnect(msg) => {},
+                CloseMsg::Normal(_msg) => {}
+                CloseMsg::Disconnect(_msg) => {}
             }
-
         })?;
 
         drop(socket_borrow);
@@ -102,94 +102,17 @@ impl Transport {
         self.subs.borrow_mut().push(sub);
     }
 
-    pub fn send_command(&self, command: &Command) {
+    pub fn send_command(&self, command: Command) {
         let socket_borrow = self.sock.borrow();
 
         // TODO: no socket? we dont really want this method to return err
         if let Some(socket) = socket_borrow.as_ref() {
-            let msg = serde_json::to_string(&command).unwrap();
-
-            socket.send(&msg).unwrap();
+            socket.send(&OutMsg::Command(command)).unwrap();
         }
     }
 }
 
-struct Pinger(Rc<RefCell<InnerPinger>>);
-
-struct InnerPinger {
-    ping_interval: i32,
-    num: usize,
-    pong_at: Option<f64>,
-    socket: Rc<RefCell<Option<WebSocket>>>,
-    ping_task: Option<PingTaskHandler>,
-}
-
-impl InnerPinger {
-    fn send_now(&mut self) -> Result<(), WasmErr> {
-        let borrow = self.socket.try_borrow()?;
-        let socket = borrow
-            .as_ref()
-            .ok_or(WasmErr::from_str("Unable to ping: no socket"))?;
-        self.num += 1;
-        let msg = serde_json::to_string(&Heartbeat::Ping(self.num))?;
-        socket.send(&msg)
-    }
-}
-
-struct PingTaskHandler {
-    _ping_closure: Closure<dyn FnMut()>,
-    _interval_handler: IntervalHandle,
-}
-
-impl Pinger {
-    fn new(ping_interval: i32) -> Self {
-        Self(Rc::new(RefCell::new(InnerPinger {
-            ping_interval,
-            num: 0,
-            pong_at: None,
-            socket: Rc::new(RefCell::new(None)),
-            ping_task: None,
-        })))
-    }
-
-    fn set_pong_at(&self, at: f64) {
-        self.0.borrow_mut().pong_at = Some(at);
-    }
-
-    fn start(
-        &self,
-        socket: Rc<RefCell<Option<WebSocket>>>,
-    ) -> Result<(), WasmErr> {
-        let mut inner = self.0.borrow_mut();
-        inner.socket = socket;
-        inner.send_now()?;
-
-        let inner_rc = Rc::clone(&self.0);
-        let do_ping = Closure::wrap(Box::new(move || {
-            // its_ok if ping fails few times
-            inner_rc.borrow_mut().send_now().is_ok();
-        }) as Box<dyn FnMut()>);
-
-        let interval_id = window()
-            .set_interval_with_callback_and_timeout_and_arguments_0(
-                do_ping.as_ref().unchecked_ref(),
-                inner.ping_interval,
-            )?;
-
-        inner.ping_task = Some(PingTaskHandler {
-            _ping_closure: do_ping,
-            _interval_handler: IntervalHandle(interval_id),
-        });
-
-        Ok(())
-    }
-
-    fn stop(&self) {
-        self.0.borrow_mut().ping_task.take();
-    }
-}
-
-enum CloseMsg {
+pub enum CloseMsg {
     Normal(String),
     Disconnect(String),
 }
